requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
faiss-cpu
youtube-transcript-api
sentence-transformers
python-dotenv
pydantic
requests
